# backend/public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'AlleyNote API'
  description: "\nAlleyNote 公布欄系統 API 文件\n\n這是一個現代化的公布欄系統 API，提供完整的貼文管理、使用者認證、檔案上傳等功能。\n\n## 認證方式\n- **JWT Bearer Token**: 使用 JWT Token 進行 API 認證\n- **Session Auth**: 傳統 Session 認證 (向下相容)\n\n## 主要功能\n- 📝 **貼文管理**: 建立、編輯、刪除、查看貼文\n- 👤 **使用者認證**: 登入、登出、註冊、密碼管理\n- 📎 **檔案上傳**: 支援附件上傳和管理\n- 🔐 **權限控制**: 基於角色的存取控制\n\n## API 版本\n當前版本: v1.0.0\n支援的回應格式: JSON\n字元編碼: UTF-8\n\n## 錯誤處理\n所有錯誤回應都遵循統一格式：\n```json\n{\n  \"success\": false,\n  \"error\": \"錯誤訊息\",\n  \"errors\": {\n    \"field\": [\"具體錯誤說明\"]\n  }\n}\n```\n\n## 聯絡資訊\n如有問題或建議，請聯繫開發團隊。\n    "
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
    description: 開發環境伺服器
  -
    url: 'https://api.alleynote.example.com'
    description: 正式環境伺服器
paths:
...
  /api/statistics/export/views/csv:
    get:
      tags:
        - statistics
      summary: '匯出文章瀏覽統計為 CSV'
      description: "匯出文章瀏覽統計為 CSV.\n\nGET /api/statistics/export/views/csv"
      operationId: 5f37e502987491a5994544c24ba921f9
      parameters:
        -
          name: post_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 'CSV 文件'
  /api/statistics/export/comprehensive/csv:
    get:
      tags:
        - statistics
      summary: '匯出綜合分析報告為 CSV'
      description: "匯出綜合分析報告為 CSV.\n\nGET /api/statistics/export/comprehensive/csv"
      operationId: cfb140be1f6b4e7c69af1a4f2cf6dd9d
      parameters:
        -
          name: post_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 'CSV 文件'
  /api/statistics/export/comprehensive/json:
    get:
      tags:
        - statistics
      summary: '匯出綜合分析報告為 JSON'
      description: "匯出為 JSON.\n\nGET /api/statistics/export/comprehensive/json"
      operationId: 12b80f8c9d0f63101971dcec97a59d4d
      parameters:
        -
          name: post_id
          in: query
          required: false
          schema:
            type: integer
        -
          name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 'JSON 文件'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: bearer
    sessionAuth:
      type: apiKey
      name: PHPSESSID
      in: cookie
...